requirements.txt
langchain-core
langchain-community
langchain-groq
langchain-openai
pydantic
python-dotenv
langchain
defusedxml
streamlit